requirements.txt
biopython>=1.67
regex>=2016.8.27
tqdm>=4.8.4
